+++ __tests__/app-5-7-preset-react.js
@@ -12,7 +12,9 @@ describe('generator-laravel-5:app', () => {
         version: '5.7.*',
         proxy: 'localhost',
         schema: 'http',
-        preset: 'react'
+        preset: 'react',
+        enableAuth: false,
+        localGit: false
       })
       .on('end', done);
   }, 1200000);